Generation of complete ox path for graph-based cell traversal
This is a WIP and does not yet avoid obstacles. Paths through cells
that were already covered are generated simply by passing through the
centroid of that cell. The drone does not return all the way to the
point of departure either.

The new algorithm can be executed by running mytest.py. test.py
remains functional and will run the old version of the algorithm.