oxpath: determine traversal endpoints by eroding polygon
This is a better method than the one used before (simply interpolating
inward from the end of each traversal by the length of one drone
radius). With the previous method, the drone could get too close to
the edge of a cell, possible causing it to drop pesticides outside of
the field. This method removes that possibility.