oxpath: shorter transition between cells
oxpath now looks ahead to find the next cell that needs to be covered
(or the return to the start point), i.e. a cell that isn't just a
crossover node linking two important nodes in the graph. It tries to
go there directly, rather than going through each of the cells
in-between the current cell and that cell one by one. This results in
significantly shorter and more intuitive paths being generated.