Optimal exhaustive path, exhaustive search
Began to write an algorithm to generate an exhaustive path with
an exhaustive search by going through each permutation of each node's
neighbors. The algorithm falls short because there is no currently
forseeable way to incoporate the part of the algorithm which backtracks
and then, 1, compare all of the generated paths, or b, determine the
shortest path as each path is generated.