Simple optimizations
Finished implementing the first stages of a simple algorithm which
spends a certain amount of time finding the possible paths for
each start node. A randomness actor is introduced to feign a simple
optimization - we shuffle the edges each time so that there is a chance
of finding multiple alternate paths.