Update decompose.py and cellgrpaher.py
Modified decompose.py to return a tuple which contains the
a list of grand cells extracted from the polygon, a list of
trapezoids extracted from the polygon, the angle at which
the best decomposition was found, and the point around which
the polygon was rotated by that angle.

These changes are also reflected in cellgrapher.py, where I have
made it so that the original functionality of cellgrapher is
applied to the list of CellTrap instances (aka the trapezoids
described above).