oxpath.py: use cells to generate oxpath and trapezoids to transition
Significant improvement to oxpath.py. The drone path is now generated
using the large cells (which may or may not be trapezoids) instead of
trapezoidal cells (i.e. we are using a boustrophedon decomposition
instead of a trapezoidal decomposition). This minimizes the length of
unnecessary transitions between cells and maximizes the surface area
in the polygon that is covered by the drone. For transitions between
cells, the graph of trapezoids is used: a pathfinding algorithm is
used each time a transition needs to be generated.

Also modified celllinker.py to only add the cells that need to be
covered to the stack, not the transitional cells. Given that we are
now calculating the transitions using the graph of trapezoids, it is
not necessary to record the transitional cells.

Finally, modified missionplanner.py and test.py to accommodate the
changes to oxpath.py and celllinker.py.