Add coordtransform.py
Allows for conversion from GPS coordinates to position in meters
relative to a specified origin point. This is necessary because the
drone radius should not be expressed in latitude/longitude distance as
it is currently (since length between longitude lines changes
depending on the latitude), and because the drone controller we will
use refers to position in meters, not in GPS coordinates.