Clean up test.py and test fields
Moved all testing code to test directory. Deleted files with
coordinates of test fields (since we no longer need to be able to read
field coordinates from a file) and added test cases in Python code in
testfields.py. The format of the test cases is the same as the JSON
format that will be used to plan missions from a node.js server.