Remove Mission.py and creation of .waypoints output file
We no longer need an output file that is readable by MAVProxy since we
are no longer using MAVProxy to control the drone. We just need an
output list of waypoints that can be communicated to the node.js
server. This commit simplifies the repository and removes the
unnecessary functionality.